Home page created, now i can do the profile page
