+++ src/Pages/Ricks/Ricks.jsx
@@ -1,21 +1,15 @@
-import React , { useContext } from 'react';
-import { AppContext } from '../../App/App.jsx'
-import { API_URL } from '../../API';
+import React , { Fragment, useContext } from 'react';
+import { AppContext } from '../../App/App.jsx';
 import { Characters } from '../../Components/Characters/Characters.jsx';
 
+
+
 function Ricks(){
-    const { info , setCharacters } = useContext(AppContext);
-    const ricks = [];
-    for( let i=1 ; i<=info.count ; i++ ){
-        fetch( `${API_URL}/${i}` )
-        .then( res => res.json())
-        .then( data => {
-            if( data.name.includes('Rick') ) ricks.push(data);
-        })
-        setCharacters(ricks);
-    }
+    const { xd  } = useContext(AppContext);
     return(
-        <Characters />
+        <Fragment>
+        { xd() && <Characters /> }
+        </Fragment>
     )
 }
